mallory/requirements.txt
mpmath==1.3.0
pycryptodome==3.19.0
sympy==1.12
